Revert "Move RDHAny from DataFormats/Headers to Detectors/Raw"
This reverts commit 833ae764b43708204825bbecbb4c57095b2e1ccb.
Since the GPUCommon was freed from root dependency, the RDHAny can stay in Headers.
