+++ DataFormats/Headers/CMakeLists.txt
@@ -11,8 +11,9 @@
 o2_add_library(Headers
                SOURCES src/DataHeader.cxx src/NameHeader.cxx
                        src/HeartbeatFrame.cxx src/TimeStamp.cxx
-                       src/DAQID.cxx
-               PUBLIC_LINK_LIBRARIES O2::MemoryResources)
+                       src/DAQID.cxx src/RDHAny.cxx
+               PUBLIC_LINK_LIBRARIES O2::MemoryResources
+                                     O2::GPUCommon)
 
 o2_add_test(DataHeader
             SOURCES test/testDataHeader.cxx
