+++ Detectors/MUON/MCH/Raw/Decoder/src/RDHManip.cxx
@@ -16,7 +16,7 @@
 #include <iostream>
 #include <string>
 #include "CommonConstants/Triggers.h"
-#include "DetectorsRaw/RDHAny.h"
+#include "Headers/RDHAny.h"
 #include "DetectorsRaw/RDHUtils.h"
 
 namespace
